requirements.txt
imutils
azure
msrest
matplotlib
PILLOW
requests
opencv-python
azure-cognitiveservices-vision-face
asyncio